feat(core): add viewport helper to state
