+++ packages/core/src/store/actions.ts
@@ -180,22 +180,6 @@ export function useActions(
       return res
     }, [])
 
-    if (!state.fitViewOnInitDone && state.fitViewOnInit) {
-      viewportHelper.value.fitView()
-    }
-
-    state.fitViewOnInitDone = true
-
-    // Here, we are making all nodes visible once we have the dimensions.
-    if (!document.querySelector('#vue-flow__initialized-styles')) {
-      const style = document.createElement('style')
-      style.id = 'vue-flow__initialized-styles'
-      document.head.appendChild(style)
-
-      const css = `.vue-flow__node { visibility: visible !important; }`
-      style.appendChild(document.createTextNode(css))
-    }
-
     if (changes.length) {
       state.hooks.nodesChange.trigger(changes)
     }
@@ -954,6 +938,7 @@ export function useActions(
     toObject,
     fromObject,
     updateNodeInternals,
+    viewportHelper,
     $reset,
     $destroy: () => {},
   }
