+++ packages/core/src/store/state.ts
@@ -51,8 +51,6 @@ function defaultState(): State {
     d3Selection: null,
     d3ZoomHandler: null,
 
-    viewportInitialized: false,
-
     minZoom: 0.5,
     maxZoom: 2,
